refactor classes to just two classes
