+++ user.txt
@@ -1,8 +1,8 @@
-My name is   wandji best name ever
+    My name is   wandji best name ever    
 I am from Cameroon
  I love coding
-cooking, music,    fun chats and indoor drinking are my hobbies   
-( the man )
+cooking, music,    fun chats and indoor drinking are my hobbies.   
+( the man.  )
 
   
-name  check 
+  name  check
